create XsCalendarUi component
